searching for incomming call in localdb before searching in server

+++ app/src/main/java/com/nibble/hashcaller/network/search/model/Cntct.kt
@@ -25,7 +25,7 @@ data class Cntct(
     val country: String?="",
 
     @SerializedName("spamCount")
-    val spammCount : Int? = 0,
+    val spammCount : Long? = 0,
